Reorganize onboarding package into welcome/ and signin/ subpackages
